+++ usecase/src/test/java/org/mobilitydata/gtfsvalidator/usecase/usecasevalidator/ShapeBasedCrossValidatorTest.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package org.mobilitydata.gtfsvalidator.usecase.usecaseprocessor;
+package org.mobilitydata.gtfsvalidator.usecase.usecasevalidator;
 
 import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.Test;
